yolo_v8_face/requirements.txt
pillow
numpy
accelerate
sentencepiece
huggingface-hub
supervision
ultralytics
openvino